+++ services/devstar/v1/model/model_list_published_templates_request.go
@@ -1,7 +1,7 @@
 /*
- * Devstar
+ * DevStar
  *
- * Devstar API
+ * DevStar API
  *
  */