+++ services/elb/v3/model/model_list_availability_zones_response.go
@@ -15,7 +15,7 @@ import (
 
 // Response Object
 type ListAvailabilityZonesResponse struct {
-	// 可用区列表
+	// 可用区列表  > 获取可用区列表后，在（如创建LB时）设置可用区，选择的多个可用区必须同时在同一个子集中
 	AvailabilityZones *[][]AvailabilityZone `json:"availability_zones,omitempty"`
 	// 请求ID。 注：自动生成
 	RequestId *string `json:"request_id,omitempty"`