+++ services/elb/v3/model/model_member_ref.go
@@ -16,7 +16,7 @@ import (
 //
 type MemberRef struct {
 	// 后端服务器ID。
-	Id *string `json:"id,omitempty"`
+	Id string `json:"id"`
 }
 
 func (o MemberRef) String() string {