+++ services/elb/v3/model/model_page_info.go
@@ -15,9 +15,9 @@ import (
 
 // 翻页信息
 type PageInfo struct {
-	// 分页查询中上页的ID
+	// 分页查询结果中第一条记录的ID
 	PreviousMarker string `json:"previous_marker"`
-	// 分页查询中下页的ID。 注：分页的场合输出该字段。
+	// 分页查询结果中最后一条记录的ID。
 	NextMarker *string `json:"next_marker,omitempty"`
 	// 当前的记录数。
 	CurrentCount int32 `json:"current_count"`