+++ services/elb/v3/model/model_update_certificate_response.go
@@ -16,8 +16,8 @@ import (
 // Response Object
 type UpdateCertificateResponse struct {
 	// 请求ID。 注：自动生成 。
-	RequestId   *string      `json:"request_id,omitempty"`
-	Certificate *Certificate `json:"certificate,omitempty"`
+	RequestId   *string          `json:"request_id,omitempty"`
+	Certificate *CertificateInfo `json:"certificate,omitempty"`
 }
 
 func (o UpdateCertificateResponse) String() string {