+++ CHANGELOG.md
@@ -1,5 +1,15 @@
+## 0.0.18-beta 2020-10-20
+## HuaweiCloud SDK ELB
+ - ### Features
+    - Support more interfaces of version v2.
+ - ### Bug Fix
+    - None
+ - ### Change
+    - None
+
+
 ## 0.0.17-beta 2020-10-14
-## HuaweiCloud SDK Bss
+## HuaweiCloud SDK BSS
  - ### Features
     - Partner center supports exporting product catalog prices.
  - ### Bug Fix
@@ -385,3 +395,13 @@
     - Identity and Access Management(IAM)
     - Tag Management Service(TMS)
     - Virtual Private Cloud(VPC)
+   
+# 3.0.17-beta 2020-10-16
+## HuaweiCloud SDK DevStar
+ - ### Features
+    - Change of title from Devstar to DevStar.
+    - Supplemental DevStar Automation Test Cases.
+ - ### Bug Fix
+    - None
+ - ### Change
+    - None